+++ src/utils/providers/lrclib.ts
@@ -10,7 +10,7 @@ export const fetchLRCLIB = async (playing: Playing): Promise<Lyric> => {
         const response: Lyric = {
                 id: playing.id,
                 provider: Provider.LRCLIB
-        }
+        };
         const url = `https://lrclib.net/api/search?artist_name=${playing.artistName?.replaceAll(
                 " ",
                 "+"
@@ -24,7 +24,7 @@ export const fetchLRCLIB = async (playing: Playing): Promise<Lyric> => {
                 response.exception = {
                         code: err.response?.status ?? 404,
                         message: err.response ? `${err.response.data}` : err.message
-                }
+                };
                 return response;
         }
         const songs = result.data;
@@ -33,23 +33,23 @@ export const fetchLRCLIB = async (playing: Playing): Promise<Lyric> => {
                 response.exception = {
                         code: 404,
                         message: 'Cannot find track'
-                }
+                };
                 return response;
         }
 
-        const closest = songs.find(song => song.syncedLyrics)
+        const closest = songs.find(song => song.syncedLyrics);
         if (closest) {
-                response.syncedLyric = closest.syncedLyrics?.split('\n')
-                response.plainLyric = closest.plainLyrics?.split('\n')
-                response.instrumental = closest.instrumental
+                response.syncedLyric = closest.syncedLyrics?.split('\n');
+                response.plainLyric = closest.plainLyrics?.split('\n');
+                response.instrumental = closest.instrumental;
         } else {
-                response.instrumental = songs[0].instrumental
+                response.instrumental = songs[0].instrumental;
         }
         if (!response.plainLyric) {
-                const plain = songs.find(song => song.plainLyrics)
+                const plain = songs.find(song => song.plainLyrics);
                 if (plain) {
-                        response.plainLyric = plain.plainLyrics?.split('\n')
-                        response.instrumental ??= plain.instrumental
+                        response.plainLyric = plain.plainLyrics?.split('\n');
+                        response.instrumental ??= plain.instrumental;
                 }
         }
         if (!response.syncedLyric && !response.plainLyric) {
@@ -60,7 +60,7 @@ export const fetchLRCLIB = async (playing: Playing): Promise<Lyric> => {
                 response.exception = {
                         code: 404,
                         message: 'No lyrics'
-                }
+                };
                 return response;
         }